Move definition for NetVis.Render() to View section
jbenet wrote that:
one thing to keep in mind while building the code for this is to keep "the simulator"
and "the animation" separate in the code, so we can swap out rendering later
(i.e. shift between "circle view" (great for understanding protocols, and for small networks),

and it makes sense to me.

So now I am trying to put all the rendering and representation specific things into the files within the view directory.

And all the controller-apropriate definitions to controller directory.