Overwrite humanTimeLabel for last timeInterval
History consists of  timeintervals, not time moments.
In order to have interval for the initial and final steps we add past and future timestamps.
Thus we need to mark the first and last timeinterval with _ends and _starts time moments.

We take the last time interval and relabel it here.

The first time interval is relabeled in the timeinterval constructor as it is easy to tell there when the initialized interval is the first one.